Add helper method
- Share logic btw two code flow
