+++ src/Features/LanguageServer/Protocol/Handler/SemanticTokens/SemanticTokensRangeHandler.cs
@@ -2,9 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.CodeAnalysis.Classification;
 using Microsoft.CodeAnalysis.Options;
 using Microsoft.VisualStudio.LanguageServer.Protocol;
 using Roslyn.Utilities;
@@ -41,32 +41,13 @@ namespace Microsoft.CodeAnalysis.LanguageServer.Handler.SemanticTokens
             CancellationToken cancellationToken)
         {
             Contract.ThrowIfNull(request.TextDocument, "TextDocument is null.");
-            var contextDocument = context.GetRequiredDocument();
 
-            // If the full compilation is not yet available, we'll try getting a partial one. It may contain inaccurate
-            // results but will speed up how quickly we can respond to the client's request.
-            var document = contextDocument.WithFrozenPartialSemantics(cancellationToken);
-            var project = document.Project;
-            var options = _globalOptions.GetClassificationOptions(project.Language) with { ForceFrozenPartialSemanticsForCrossProcessOperations = true };
-
-            // The results from the range handler should not be cached since we don't want to cache
-            // partial token results. In addition, a range request is only ever called with a whole
-            // document request, so caching range results is unnecessary since the whole document
-            // handler will cache the results anyway.
-            var capabilities = context.GetRequiredClientCapabilities();
             var ranges = new LSP.Range[1] { request.Range };
-            var tokensData = await SemanticTokensHelpers.ComputeSemanticTokensDataAsync(
-                capabilities,
-                document,
-                ranges,
-                options,
-                cancellationToken).ConfigureAwait(false);
-
-            // The above call to get semantic tokens may be inaccurate (because we use frozen partial semantics).  Kick
-            // off a request to ensure that the OOP side gets a fully up to compilation for this project.  Once it does
-            // we can optionally choose to notify our caller to do a refresh if we computed a compilation for a new
-            // solution snapshot.
-            await _semanticTokenRefreshQueue.TryEnqueueRefreshComputationAsync(project, cancellationToken).ConfigureAwait(false);
+            var tokensData = await SemanticTokensHelpers.HandleRequestHelperAsync(_globalOptions, _semanticTokenRefreshQueue, ranges, context, cancellationToken).ConfigureAwait(false);
+            if (tokensData is null)
+            {
+                return new LSP.SemanticTokens() { Data = Array.Empty<int>() };
+            }
 
             return new LSP.SemanticTokens { Data = tokensData };
         }
